refactor: like mutation separated
